+++ src/main/java/at/hannibal2/skyhanni/utils/APIUtil.kt
@@ -45,7 +45,7 @@ object APIUtil {
                         return parser.parse(retSrc) as JsonObject
                     } catch (e: JsonSyntaxException) {
                         if (retSrc.contains("<center><h1>502 Bad Gateway</h1></center>")) {
-                            LorenzUtils.error("[SkyHanni] HyPixel API is down :(")
+                            LorenzUtils.error("[SkyHanni] Hypixel API is down :(")
                         } else {
                             println("JsonSyntaxException at getJSONResponse '$urlString'")
                             LorenzUtils.error("[SkyHanni] JsonSyntaxException at getJSONResponse!")